requirements.txt
streamlit==1.32.0
openai==1.14.0
pdfplumber==0.10.3
python-dotenv==1.0.1
reportlab==4.1.0
requests==2.31.0
Pillow==10.2.0
pytest==8.1.1
fastapi==0.110.0
uvicorn[standard]==0.27.1
